Netologia_Django_3_4/requirements-dev.txt
pytest
pytest-django
djangorestframework
model_bakery
